requirements.txt
asgiref==3.7.2
attrs==23.1.0
colorama==0.4.6
Django==4.2.6
django-environ==0.11.2
djangorestframework==3.14.0
drf-spectacular==0.26.5
Faker==19.12.0
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
packaging==23.2
Pillow==10.1.0
pluggy==1.3.0
psycopg2==2.9.9
pytest==7.4.3
pytest-django==4.7.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.31.0
rpds-py==0.13.1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
